Merge Swerve Start (#2)
* Nov5 Architecture Update

* New Swerve Kit Cooking

Not sure if this will work, I need to add the angular power assignments and the absolute encoder integration, but once i've done that, all should be good

* Overhaul + Angular setup added

Old system was pretty messy and chunky to read or debug.  Now things are definitely more spread out, but it's more concise.  PIDF Controller for the Angular calculations is set up.  Next steps: commenting, and a simulator in Python for VSC

* This is the NEWEST GRADLE VERSION, USE!!!!

* Gradle is fucking weird, but this is the new Gradle version that needs to be pushed

* Road Runner Settings Adapted

* workin on swerve auto

* Servo travel minimization

- Fixed getPosition() method, to give the correct value
- Wrote an algorithm to calculate the angle to rotate the servo to in order to minimize the travel distance, and therefore the time spent moving the servo.

NOTE: We will need to figure out a way of keeping track of when the motor should be reversed, and when it shouldn't. A separate variable in the SwerveServo class that denotes motor direction, perhaps?

* comm

* SwerveModule,  comments on code, initialization

SwerveModule has some reduced computing time by removing unnecessary computation.

Emerson, I left you some comments, I think all is good, but we can talk on Thursday or on slack or something.

X and Y are properly initialized.

* Bizarre Encoder bug

Should be fixed now, will test tonight.

* abs enc trouble

added localizer/testing class

* addding to swerve auto

* PID work + re-organization of power application

weird things going on with the swerve drivetrain.  I reorganized the way that i send power to the servos, but overall, no idea what the hell is going on.  also set the PIDF to 0,0,0,0 just to see if that does anything.  Added class to test just servos.

* angles not working, motor is responding to pid values

* toiling away

* Kotlin Auto Classes

* i found code to steal, we will revisit this

* Kinematics Finished, removal of RoadRunner 1.0

* Add Roadrunner 0.5.6

* Moving back to 1.0

* Reorganization for Merge Request

---------

Co-authored-by: esmersonian216 <120067536+esmersonian216@users.noreply.github.com>